OtherAuthOptionsPage overflows on small screens Fixes #178
