+++ test/widget/auth/other_auth_options_page_widget_test.dart
@@ -2,12 +2,13 @@ import 'package:crowdleague/actions/auth/sign_in_with_apple.dart';
 import 'package:crowdleague/actions/auth/sign_in_with_email.dart';
 import 'package:crowdleague/actions/auth/sign_in_with_google.dart';
 import 'package:crowdleague/actions/auth/sign_up_with_email.dart';
-import 'package:crowdleague/actions/auth/update_other_auth_options_page.dart';
+import 'package:crowdleague/actions/auth/update_email_auth_options_page.dart';
 import 'package:crowdleague/enums/auth_step.dart';
 import 'package:crowdleague/enums/device/platform_type.dart';
 import 'package:crowdleague/enums/email_auth_mode.dart';
 import 'package:crowdleague/models/app/app_state.dart';
 import 'package:crowdleague/reducers/app_reducer.dart';
+import 'package:crowdleague/widgets/auth/email_auth_options_page.dart';
 import 'package:crowdleague/widgets/auth/other_auth_options_page.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_auth_buttons/flutter_auth_buttons.dart';
@@ -145,8 +146,7 @@ void main() {
       expect(testMiddleware.received(SignInWithApple()), true);
     });
 
-    testWidgets(
-        'shows sign in UI when user lands on page, and after user taps sign in chip',
+    testWidgets('shows sign in UI when user taps sign in chip',
         (WidgetTester tester) async {
       // Setup the app state with expected values
       final initialAppState = AppState.init();
@@ -155,7 +155,7 @@ void main() {
 
       // Create the test harness.
       final store = Store<AppState>(appReducer, initialState: alteredState);
-      final wut = OtherAuthOptionsPage();
+      final wut = EmailAuthOptionsPage();
       final harness =
           StoreProvider<AppState>(store: store, child: MaterialApp(home: wut));
 
@@ -176,7 +176,7 @@ void main() {
       expect(signInButton, findsOneWidget);
     });
 
-    testWidgets('shows create account UI after user taps create account chip',
+    testWidgets('shows create account UI when user taps create account chip',
         (WidgetTester tester) async {
       // Setup the app state with expected values
       final initialAppState = AppState.init();
@@ -185,7 +185,7 @@ void main() {
 
       // Create the test harness.
       final store = Store<AppState>(appReducer, initialState: alteredState);
-      final wut = OtherAuthOptionsPage();
+      final wut = EmailAuthOptionsPage();
       final harness =
           StoreProvider<AppState>(store: store, child: MaterialApp(home: wut));
 
@@ -203,7 +203,7 @@ void main() {
     });
 
     testWidgets(
-        'dispatches UpdateOtherAuthOptionsPage(EmailAuthMode.signUp) action when tap sign in with email chip',
+        'dispatches UpdateEmailAuthOptionsPage(EmailAuthMode.signUp) on sign in with email chip',
         (WidgetTester tester) async {
       // Setup the app state with expected values
       final initialAppState = AppState.init();
@@ -212,7 +212,7 @@ void main() {
       // Create the test harness.
       final store = Store<AppState>(appReducer,
           initialState: initialAppState, middleware: [testMiddleware]);
-      final wut = OtherAuthOptionsPage();
+      final wut = EmailAuthOptionsPage();
       final harness =
           StoreProvider<AppState>(store: store, child: MaterialApp(home: wut));
 
@@ -229,12 +229,12 @@ void main() {
       // check correct action is dispatched
       expect(
           testMiddleware
-              .received(UpdateOtherAuthOptionsPage(mode: EmailAuthMode.signUp)),
+              .received(UpdateEmailAuthOptionsPage(mode: EmailAuthMode.signUp)),
           true);
     });
 
     testWidgets(
-        'dispatches UpdateOtherAuthOptionsPage(EmailAuthMode.signIn) action when tap create an account chip',
+        'dispatches UpdateEmailAuthOptionsPage(EmailAuthMode.signIn) on create an account chip',
         (WidgetTester tester) async {
       // Setup the app state with expected values
       final initialAppState = AppState.init();
@@ -243,7 +243,7 @@ void main() {
       // Create the test harness.
       final store = Store<AppState>(appReducer,
           initialState: initialAppState, middleware: [testMiddleware]);
-      final wut = OtherAuthOptionsPage();
+      final wut = EmailAuthOptionsPage();
       final harness =
           StoreProvider<AppState>(store: store, child: MaterialApp(home: wut));
 
@@ -260,11 +260,11 @@ void main() {
       // check correct action is dispatched
       expect(
           testMiddleware
-              .received(UpdateOtherAuthOptionsPage(mode: EmailAuthMode.signIn)),
+              .received(UpdateEmailAuthOptionsPage(mode: EmailAuthMode.signIn)),
           true);
     });
 
-    testWidgets('EmailTextField dispatches UpdateOtherAuthOptionsPage action ',
+    testWidgets('EmailTextField dispatches UpdateEmailAuthOptionsPage',
         (WidgetTester tester) async {
       // Setup the app state with expected values
       final initialAppState = AppState.init();
@@ -273,7 +273,7 @@ void main() {
       // Create the test harness.
       final store = Store<AppState>(appReducer,
           initialState: initialAppState, middleware: [testMiddleware]);
-      final wut = OtherAuthOptionsPage();
+      final wut = EmailAuthOptionsPage();
       final harness =
           StoreProvider<AppState>(store: store, child: MaterialApp(home: wut));
 
@@ -290,12 +290,11 @@ void main() {
 
       // check correct action is dispatched
       expect(
-          testMiddleware.received(UpdateOtherAuthOptionsPage(email: testEmail)),
+          testMiddleware.received(UpdateEmailAuthOptionsPage(email: testEmail)),
           true);
     });
 
-    testWidgets(
-        'PasswordTextField dispatches UpdateOtherAuthOptionsPage action',
+    testWidgets('PasswordTextField dispatches UpdateEmailAuthOptionsPage',
         (WidgetTester tester) async {
       // Setup the app state with expected values
       final initialAppState = AppState.init();
@@ -304,7 +303,7 @@ void main() {
       // Create the test harness.
       final store = Store<AppState>(appReducer,
           initialState: initialAppState, middleware: [testMiddleware]);
-      final wut = OtherAuthOptionsPage();
+      final wut = EmailAuthOptionsPage();
       final harness =
           StoreProvider<AppState>(store: store, child: MaterialApp(home: wut));
 
@@ -322,12 +321,11 @@ void main() {
       // check correct action is dispatched
       expect(
           testMiddleware
-              .received(UpdateOtherAuthOptionsPage(password: testPassword)),
+              .received(UpdateEmailAuthOptionsPage(password: testPassword)),
           true);
     });
 
-    testWidgets(
-        'RepeatPasswordTextField dispatches UpdateOtherAuthOptionsPage action',
+    testWidgets('RepeatPasswordTextField dispatches UpdateEmailAuthOptionsPage',
         (WidgetTester tester) async {
       // Setup the app state with expected values
       final initialAppState = AppState.init();
@@ -338,7 +336,7 @@ void main() {
       // Create the test harness.
       final store = Store<AppState>(appReducer,
           initialState: alteredState, middleware: [testMiddleware]);
-      final wut = OtherAuthOptionsPage();
+      final wut = EmailAuthOptionsPage();
       final harness =
           StoreProvider<AppState>(store: store, child: MaterialApp(home: wut));
 
@@ -358,11 +356,11 @@ void main() {
       // check correct action is dispatched
       expect(
           testMiddleware.received(
-              UpdateOtherAuthOptionsPage(repeatPassword: testPassword)),
+              UpdateEmailAuthOptionsPage(repeatPassword: testPassword)),
           true);
     });
 
-    testWidgets('sign in button dispatches SignInWithEmail action',
+    testWidgets('sign in button dispatches SignInWithEmail',
         (WidgetTester tester) async {
       // Setup the app state with expected values
       final initialAppState = AppState.init();
@@ -371,7 +369,7 @@ void main() {
       // Create the test harness.
       final store = Store<AppState>(appReducer,
           initialState: initialAppState, middleware: [testMiddleware]);
-      final wut = OtherAuthOptionsPage();
+      final wut = EmailAuthOptionsPage();
       final harness =
           StoreProvider<AppState>(store: store, child: MaterialApp(home: wut));
 
@@ -389,7 +387,7 @@ void main() {
       expect(testMiddleware.received(SignInWithEmail()), true);
     });
 
-    testWidgets('create account button dispatches SignUpWithEmail action',
+    testWidgets('create account button dispatches SignUpWithEmail',
         (WidgetTester tester) async {
       // Setup the app state with expected values
       final initialAppState = AppState.init();
@@ -400,7 +398,7 @@ void main() {
       // Create the test harness.
       final store = Store<AppState>(appReducer,
           initialState: alteredState, middleware: [testMiddleware]);
-      final wut = OtherAuthOptionsPage();
+      final wut = EmailAuthOptionsPage();
       final harness =
           StoreProvider<AppState>(store: store, child: MaterialApp(home: wut));
 
@@ -428,7 +426,7 @@ void main() {
       // Create the test harness.
       final store = Store<AppState>(appReducer,
           initialState: alteredState, middleware: [testMiddleware]);
-      final wut = OtherAuthOptionsPage();
+      final wut = EmailAuthOptionsPage();
       final harness =
           StoreProvider<AppState>(store: store, child: MaterialApp(home: wut));
